+++ enki/blfs/copy-to-share.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+cp -r config/* /tmp/vm-share
+cp -r scripts/* /tmp/vm-share